New module - Social_Post_Logger
With this addition, I have a standardized class that will keep track of posting for each social media site.  
This will keep track of each post, and report back if a new post is "allowed" to take place.  It keeps track of x number of post for the previous 24 hour period.